🐛 fix: resolve CI failures for PR #21857
- Move CopyButton file's extractCodeBlocks helper out of MissionDetailView.parts.tsx
  into a new MissionDetailView.helpers.ts to satisfy react-refresh/only-export-components.
- Add eslint-disable-next-line no-restricted-syntax comments (matching existing
  precedent in ConnectTab.tsx) for raw checkbox <input> elements in
  ResolutionHistoryPanel.parts.tsx and StandaloneOrbitDialog.parts.tsx, since no
  <Checkbox> component exists yet.
- Update .eslint-baseline.json entries for rbac/CanIChecker.tsx -> CanIChecker.parts.tsx
  to reflect the file split from PR #21839 that had left the baseline stale, which
  was causing these pre-existing violations to be misreported as new.

Other failing checks (Analyze Go, Analyze JavaScript/TypeScript CodeQL, detect-untested)
fail due to 'API rate limit exceeded for installation' errors from GitHub's API and are
unrelated to this PR's code changes; they should resolve automatically on rerun.
