+++ web/.eslint-baseline.json
@@ -6587,24 +6587,24 @@
     "message": "'React' is defined but never used. Allowed unused vars must match /^_/u."
   },
   {
-    "file": "src/components/rbac/CanIChecker.tsx",
+    "file": "src/components/rbac/CanIChecker.parts.tsx",
     "rule": "no-restricted-syntax",
-    "line": 264,
-    "column": 13,
+    "line": 33,
+    "column": 9,
     "message": "Use <Select> from components/ui/Select.tsx instead of raw <select>."
   },
   {
-    "file": "src/components/rbac/CanIChecker.tsx",
+    "file": "src/components/rbac/CanIChecker.parts.tsx",
     "rule": "no-restricted-syntax",
-    "line": 285,
-    "column": 13,
+    "line": 117,
+    "column": 9,
     "message": "Use <Select> from components/ui/Select.tsx instead of raw <select>."
   },
   {
-    "file": "src/components/rbac/CanIChecker.tsx",
+    "file": "src/components/rbac/CanIChecker.parts.tsx",
     "rule": "no-restricted-syntax",
-    "line": 300,
-    "column": 13,
+    "line": 58,
+    "column": 5,
     "message": "Use <Input> from components/ui/Input.tsx instead of raw <input>."
   },
   {
@@ -6615,10 +6615,10 @@
     "message": "Use <Select> from components/ui/Select.tsx instead of raw <select>."
   },
   {
-    "file": "src/components/rbac/CanIChecker.tsx",
+    "file": "src/components/rbac/CanIChecker.parts.tsx",
     "rule": "no-restricted-syntax",
-    "line": 332,
-    "column": 13,
+    "line": 137,
+    "column": 9,
     "message": "Use <Input> from components/ui/Input.tsx instead of raw <input>."
   },
   {
